refactor: merge _customfield option files into existing option files
Move all custom-field-specific option builders into their corresponding
type-grouped files and delete the _customfield split files:

- option_bool.go: add WithAllowAddItem, WithAllowInput (list type)
- option_int.go: add WithInitialShift, WithInitialValueType (date type)
- option_float.go: new file for WithMin, WithMax, WithInitialValue (number type)
- option_string.go: add WithUnit, WithInitialDateMin, WithInitialDateMax,
  WithInitialDate (number/date type)
- option_slice.go: add WithItems (list type)
- remove option_bool_customfield.go, option_int_customfield.go,
  option_float_customfield.go, option_string_customfield.go,
  option_slice_customfield.go
