--- internal/core/option_float_customfield.go
@@ -1,25 +1,0 @@
-package core
-
-// WithMin returns an option to set the `min` parameter for Number type custom fields.
-func (s *OptionService) WithMin(min float64) RequestOption {
-	return &APIParamOption{
-		Type:    ParamMin,
-		SetFunc: setFloat64Func(ParamMin, min),
-	}
-}
-
-// WithMax returns an option to set the `max` parameter for Number type custom fields.
-func (s *OptionService) WithMax(max float64) RequestOption {
-	return &APIParamOption{
-		Type:    ParamMax,
-		SetFunc: setFloat64Func(ParamMax, max),
-	}
-}
-
-// WithInitialValue returns an option to set the `initialValue` parameter for Number type custom fields.
-func (s *OptionService) WithInitialValue(initialValue float64) RequestOption {
-	return &APIParamOption{
-		Type:    ParamInitialValue,
-		SetFunc: setFloat64Func(ParamInitialValue, initialValue),
-	}
-}
