+++ internal/core/option_string.go
@@ -47,6 +47,33 @@ func (s *OptionService) WithHookURL(hookURL string) RequestOption {
 	return nonEmptyStringOption(ParamHookURL, hookURL)
 }
 
+// WithInitialDate returns an option to set the `initialDate` parameter for Date type custom fields.
+// The value must be formatted as "yyyy-MM-dd".
+func (s *OptionService) WithInitialDate(date string) RequestOption {
+	return &APIParamOption{
+		Type:    ParamInitialDate,
+		SetFunc: setStringFunc(ParamInitialDate, date),
+	}
+}
+
+// WithInitialDateMax returns an option to set the `max` parameter for Date type custom fields.
+// The value must be formatted as "yyyy-MM-dd".
+func (s *OptionService) WithInitialDateMax(date string) RequestOption {
+	return &APIParamOption{
+		Type:    ParamMax,
+		SetFunc: setStringFunc(ParamMax, date),
+	}
+}
+
+// WithInitialDateMin returns an option to set the `min` parameter for Date type custom fields.
+// The value must be formatted as "yyyy-MM-dd".
+func (s *OptionService) WithInitialDateMin(date string) RequestOption {
+	return &APIParamOption{
+		Type:    ParamMin,
+		SetFunc: setStringFunc(ParamMin, date),
+	}
+}
+
 // WithKey returns a option that sets the `key` field.
 func (s *OptionService) WithKey(key string) RequestOption {
 	return nonEmptyStringOption(ParamKey, key)
@@ -158,5 +185,13 @@ func (s *OptionService) WithTextFormattingRule(format model.Format) RequestOptio
 			return nil
 		},
 		SetFunc: setStringFunc(ParamTextFormattingRule, string(format)),
+	}
+}
+
+// WithUnit returns an option to set the `unit` parameter for Number type custom fields.
+func (s *OptionService) WithUnit(unit string) RequestOption {
+	return &APIParamOption{
+		Type:    ParamUnit,
+		SetFunc: setStringFunc(ParamUnit, unit),
 	}
 }
